The player character's soundset is now selectable.
WeiDU writes into the CHARSND.2DA, which now makes it possible to select the specific soundset during character creation.
In addition, a sound sample is included, which is not always played accordingly.
Also, the sound being played is not printed in the console.
All of this requires further testing as the reasons are not clear.