The soundset will now play correctly and the text will be printed to the console.
The only difference is that the sound and text are now checked using the soundset name using regex.
Additionally, there is an if statement that ensures that similar soundset names do not conflict with the soundset texts and sounds.

Additionally, the soundset was installed using the soundset tool. After uninstalling and installing with this system, the sound sets worked.
This was the case even after copying a backup and overwriting the game files.
Further testing is required for a more precise understanding.