Fixes issue #4
Audacity version 3.7.3 is so fundamental that the backup files saved in this version are no longer backward compatible.
For this reason, the manuals (English and German) have been significantly changed.
To keep maintenance work manageable, the manuals in the wiki will no longer be updated; instead, the PDFs will be updated.